Add files via upload
This Git repository contains an Arduino sketch for interfacing with the BNO055 and  Lily go T watch sensor and transmitting its orientation, gyroscope, and acceleration data over WiFi using the ESP32's built-in WiFi module. The project is designed to work with a Python listener that can receive and process the sensor data.